+++ example/main.dart
@@ -15,10 +15,12 @@ class MyApp extends StatelessWidget {
           child: FormatLabel(
             mainText: 'Animals:',
             description: 'Dog, cat, duck, hippo, bird, tiger',
-            mainTextStyle: TextStyle(
-                fontSize: 14, fontWeight: FontWeight.w700, color: Colors.red),
-            descriptionTextStyle: TextStyle(
-                fontSize: 14, fontWeight: FontWeight.w300, color: Colors.teal),
+            mainTextStyle: TextStyle(fontSize: 14, color: Colors.red),
+            descriptionTextStyle: TextStyle(fontSize: 14, color: Colors.teal),
+
+            // Optional values. Default values are bold and normal respectively.
+            mainTextWeight: FontWeight.w700,
+            descriptionTextWeight: FontWeight.w300,
           ),
         ),
       ),